// src/types/callbacks.ts
export type Callbacks = {
  readonly onDestroy?: (element: HTMLElement) => void;
  readonly onDeviceOrientation?: (element: HTMLElement) => void;
  readonly onInit?: (element: HTMLElement) => void;
  readonly onMouseEnter?: (element: HTMLElement) => void;
  readonly onMouseLeave?: (element: HTMLElement) => void;
  readonly onMouseMove?: (element: HTMLElement) => void;
};